requirements.txt
streamlit
pyrebase4
setuptools